Plotting for downsampled data
Note, if you want to run this locally the import for average_downsampling_sliding_window (something I wrote), will break since I suck and have the files laid out differently on my machine (will fix this soon). 

The downsampling works excellently! It does exactly what you expect the large the sliding window the more data is represented with each point, but the shape of the plots is maintained pretty well.